Add argument parsing

+++ bitmap.h
@@ -8,6 +8,6 @@
 
 #define pad_to_four(n) (((n + 3) / 4) * 4)
 
-void generate_bitmap(size_t img_length, size_t img_width, particle_grid_t p, size_t num_particles, const uint8_t *bg_color, const uint8_t *pt_color, char *file_name);
+void generate_bitmap(unsigned int img_length, unsigned int img_width, particle_grid_t p, size_t num_particles, const uint8_t *bg_color, const uint8_t *pt_color, char *file_name);
 
 #endif // BITMAP_H
